update post crud

+++ src/main/java/com/wiki/dev/service/PostService.java
@@ -5,24 +5,24 @@ import com.wiki.dev.dto.PostResponse;
 import com.wiki.dev.entity.Category;
 import com.wiki.dev.entity.Post;
 import com.wiki.dev.entity.User;
-import com.wiki.dev.exception.CategoryNotFoundException;
-import com.wiki.dev.exception.DevWikiException;
-import com.wiki.dev.exception.PostNotFoundException;
 import com.wiki.dev.mapper.PostMapper;
 import com.wiki.dev.repository.CategoryRepository;
 import com.wiki.dev.repository.PostRepository;
 import com.wiki.dev.repository.UserRepository;
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 @Slf4j
 @Transactional
 public class PostService {
@@ -34,49 +34,136 @@ public class PostService {
     private final PostMapper postMapper;
 
     @Transactional(readOnly = true)
-    public PostResponse getPost(Long id) {
-        Post post = postRepository.findById(id).orElseThrow(() -> new PostNotFoundException("can't find : " + id));
-        return postMapper.mapToDto(post);
+    public ResponseEntity<Map<Object, Object>> getPostById(Long id) {
+        Map<Object, Object> responseBody = new HashMap<>();
+
+        Optional<Post> optionalPost = postRepository.findById(id);
+
+        if (optionalPost.isEmpty()) {
+            responseBody.put("data", null);
+            responseBody.put("error", "optionalPost not exist !");
+            return ResponseEntity.badRequest().body(responseBody);
+        }
+
+        PostResponse postResponse = (PostResponse) optionalPost.stream().map(postMapper::mapPostToDto);
+
+        responseBody.put("data", postResponse);
+        responseBody.put("error", null);
+
+        return ResponseEntity.ok().body(responseBody);
     }
 
     @Transactional(readOnly = true)
-    public List<PostResponse> getAllPosts() {
-        return postRepository.findAll().stream().map(postMapper::mapToDto).collect(Collectors.toList());
+    public ResponseEntity<Map<Object, Object>> getAllPosts() {
+        Map<Object, Object> responseBody = new HashMap<>();
+
+        List<Post> posts = postRepository.findAll();
+
+        if (posts.isEmpty()) {
+            responseBody.put("data", "post is empty, you need to add post at least one !");
+            responseBody.put("error", null);
+            return ResponseEntity.ok().body(responseBody);
+        }
+
+        List<PostResponse> tempPosts = posts.stream().map(postMapper::mapPostToDto).collect(Collectors.toList());
+
+        responseBody.put("data", tempPosts);
+        responseBody.put("error", null);
+
+        return ResponseEntity.ok().body(responseBody);
     }
 
     @Transactional
-    public Post save(PostRequest postRequest) {
-        Category category = categoryRepository.findByName(postRequest.getCategoryName()).orElseThrow(() -> new CategoryNotFoundException("can't find : " + postRequest.getCategoryName()));
+    public ResponseEntity<Map<Object, Object>> createPost(PostRequest postRequest) {
+        Map<Object, Object> responseBody = new HashMap<>();
+
+        Optional<Category> category = categoryRepository.findByName(postRequest.getCategoryName());
+
+        if (category.isEmpty()) {
+            responseBody.put("data", null);
+            responseBody.put("error", "category not exist !");
+            return ResponseEntity.badRequest().body(responseBody);
+        }
+
         User user = authService.getCurrentUser();
+        Category tempCategory = category.get();
 
-        return postRepository.save(postMapper.map(postRequest, category, user));
+        Post post = postRepository.save(postMapper.mapToPost(postRequest, tempCategory, user));
+
+        responseBody.put("data", post);
+        responseBody.put("error", null);
+
+        return ResponseEntity.ok().body(responseBody);
     }
 
     @Transactional(readOnly = true)
-    public List<PostResponse> getPostsByCategory(Long categoryId) {
-        Category category = categoryRepository.findById(categoryId).orElseThrow(() -> new CategoryNotFoundException("can't find : " + categoryId.toString()));
-        List<Post> posts = postRepository.findAllByCategory(category);
-        return posts.stream().map(postMapper::mapToDto).collect(Collectors.toList());
+    public ResponseEntity<Map<Object, Object>> getPostsByCategory(Long categoryId) {
+        Map<Object, Object> responseBody = new HashMap<>();
+
+        Optional<Category> category = categoryRepository.findById(categoryId);
+
+        if (category.isEmpty()) {
+            responseBody.put("data", null);
+            responseBody.put("error", "category not exist !");
+            return ResponseEntity.badRequest().body(responseBody);
+        }
+
+        List<PostResponse> postResponses = postRepository.findAllByCategory(category.get())
+                .stream().map(postMapper::mapPostToDto).collect(Collectors.toList());
+
+        responseBody.put("data", postResponses);
+        responseBody.put("error", null);
+
+        return ResponseEntity.ok().body(responseBody);
     }
 
     @Transactional(readOnly = true)
-    public List<PostResponse> getPostsByUsername(String username) {
-        User user = userRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("can't find : " + username));
-        return postRepository.findByUser(user).stream().map(postMapper::mapToDto).collect(Collectors.toList());
+    public ResponseEntity<Map<Object, Object>> getPostsByUsername(String username) {
+        Optional<User> user = userRepository.findByUsername(username);
+
+        Map<Object, Object> responseBody = new HashMap<>();
+
+        if (user.isEmpty()) {
+            responseBody.put("data", null);
+            responseBody.put("error", "user not exist !");
+            return ResponseEntity.badRequest().body(responseBody);
+        }
+
+        List<PostResponse> postResponses = postRepository.findByUser(user.get())
+                .stream().map(postMapper::mapPostToDto).collect(Collectors.toList());
+
+        responseBody.put("data", postResponses);
+        responseBody.put("error", null);
+
+        return ResponseEntity.ok().body(responseBody);
     }
 
     @Transactional
-    public PostResponse updatePostById(PostRequest postRequest) {
-        Long id = postRequest.getPostId();
-        Post post = postRepository.findById(id).orElseThrow(() -> new PostNotFoundException("can't find post with : " + id));
+    public ResponseEntity<Map<Object, Object>> updatePostById(PostRequest postRequest) {
+        Map<Object, Object> responseBody = new HashMap<>();
 
-        if (postRequest.getPostName() != null) post.setPostName(postRequest.getPostName());
+        Optional<Post> optionalPost = postRepository.findById(postRequest.getPostId());
 
-        if (postRequest.getCategoryName() != null) {
-            String categoryName = postRequest.getCategoryName();
-            Category category = categoryRepository.findByName(categoryName).orElseThrow(() -> new CategoryNotFoundException("can't find category with : " + categoryName));
+        if (optionalPost.isEmpty()) {
+            responseBody.put("data", null);
+            responseBody.put("error", "post not exist !");
+            return ResponseEntity.badRequest().body(responseBody);
+        }
 
-            post.setCategory(category);
+        Post post = optionalPost.get();
+
+        if (postRequest.getPostName() == null) {
+            responseBody.put("data", null);
+            responseBody.put("error", "post name cannot be null !");
+            return ResponseEntity.badRequest().body(responseBody);
+        }
+
+        post.setPostName(postRequest.getPostName());
+
+        if (postRequest.getCategoryName() == null) {
+            responseBody.put("data", null);
+            responseBody.put("error", "you must set category for your post !");
+            return ResponseEntity.badRequest().body(responseBody);
         }
 
         if (postRequest.getUrl() != null) post.setUrl(postRequest.getUrl());
@@ -85,14 +172,38 @@ public class PostService {
 
         User user = authService.getCurrentUser();
         if (!user.getEmail().equals(post.getUser().getEmail())) {
-            throw new DevWikiException("글 작성자만 수정할 수 있습니다.");
+            responseBody.put("data", null);
+            responseBody.put("error", "post only can be updated by creator !");
+            return ResponseEntity.badRequest().body(responseBody);
         }
 
-        return postMapper.mapToDto(post);
+        post.setUser(user);
+
+        PostResponse postResponse = postMapper.mapPostToDto(post);
+
+        responseBody.put("data", postResponse);
+        responseBody.put("error", null);
+
+        return ResponseEntity.ok().body(responseBody);
     }
 
     @Transactional
-    public void deletePostById(Long id) {
+    public ResponseEntity<Map<Object, Object>> deletePostById(Long id) {
+        Map<Object, Object> responseBody = new HashMap<>();
+
+        Optional<Post> optionalPost = postRepository.findById(id);
+
+        if (optionalPost.isEmpty()) {
+            responseBody.put("data", null);
+            responseBody.put("error", "post not exist !");
+            return ResponseEntity.badRequest().body(responseBody);
+        }
+
         postRepository.deleteById(id);
+
+        responseBody.put("data", optionalPost.get().getPostName() + " is deleted !");
+        responseBody.put("error", null);
+
+        return ResponseEntity.ok().body(responseBody);
     }
 }
